Revert "change: now using cuda 9.1"
This reverts commit 56c197f59872d843caf5ee46851835ae4191c3b9.

+++ Camera/include/frame_desc.hh
@@ -37,7 +37,7 @@ namespace camera
 
 		unsigned short		width;		//!< Width of the frame in pixels.
 		unsigned short		height;		//!< Height of the frame in pixels.
-		unsigned int		depth;		//!< Byte depth during acquisition.
+		size_t				depth;		//!< Byte depth during acquisition.
 		Endianness			byteEndian;	//!< To each camera software its endianness. Useful for 16-bit cameras.
 	};
 }
